Fixed some import after moving test into test_files folder
