+++ tests/integration/test_fim/test_files/test_follow_symbolic_link/test_symlink_to_dir_between_scans.py
@@ -14,7 +14,7 @@ from wazuh_testing.tools import PREFIX
 from wazuh_testing.tools.configuration import load_wazuh_configurations, check_apply_test
 from wazuh_testing.tools.monitoring import FileMonitor
 
-from test_fim.test_follow_symbolic_link.common import wait_for_symlink_check, symlink_interval, testdir_link, \
+from test_fim.test_files.test_follow_symbolic_link.common import wait_for_symlink_check, symlink_interval, testdir_link, \
     testdir_target
